pyproject.toml
[tool.poetry]
name = "triplet-taste"
version = "0.1.0"
description = "A deep learning project for food image similarity using triplet loss and CLIP embeddings"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"
package-mode=false

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.6.0"
torchvision = "^0.21.0"
pytorch-lightning = "^2.0.0"
optuna = "^3.3.0"
pillow = "^10.0.0"
numpy = "^1.24.0"
click = "^8.1.0"
rich = "^13.0.0"
matplotlib = "^3.7.0"
scikit-learn = "^1.3.0"
requests = "^2.31.0"
psutil = "^5.9.0"
tqdm = "^4.67.1"
seaborn = "^0.13.2"
jupyter = "^1.1.1"
plotly = "^6.2.0"
kaleido = "0.2.1"
openai-clip = "^1.0.0"
transformers = "^4.55.0"
pandas = "^2.3.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"
pre-commit = "^3.4.0"



[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 